Remove `messages` from `Policy`, add `AssertError`
This commit will add code that attempts to construct a sensible error
message from the message provided to an `assert`. This is implemented by
populating the message of `AssertError`, which allows the policy pack
gRPC server to re-throw any errors that are not `instanceof` that
specific type, which allows us to incidentally resolve #44.

Fixes #23.
Fixes #44.