Add unit tests for PB serialization
In addition to adding simple unit tests, we change the gRPC server code
to correctly report when an error is thrown in an RPC call. gRPC will
hang indefinitely if you throw an exception, so here we catch all
errors, and attempt to turn them into useful gRPC error messages.