Throw when a field is unknown
When a policy is run during preview, it may access properties that can't
be known until update. This commit will cause unknown property accesses
to result in throwing an `UnknownValueError`. This allows the policy
plugin to catch unknown value accesses and report to the user that the
policy can't complete during preview, and hence may fail at runtime.

Fixes #63.