Remove the `asTypedResource` helper
This helper is no longer generally useful. For `validateResource` we want the "input" shape, but for `validateStack`, we want the "output" shape.

It'd be best to simplify things here by removing this. For `validateResource`, if you want to write your validation function in a strongly-typed way, use `validateTypedResource`. For `validateStack`, if you want to filter/cast resources, use `@pulumi/query`.