Add way to do strong typing for atypical resources
The vast majority of resources follow the same order of constructor
parameters, which allows us to infer the args type (input shape) used
when creating the resource. However, there are resources that don't
follow the pattern of constructor parameters. This change provides a way
to still have stronglyt-typed validation checks for such resources, you
just have to specify the types explicitly.