Inline `asyncTest` utility function
We recently made a change in `@pulumi/pulumi` to mark this internal helper as `@internal` so that it would not be emitted in API docs. But TypeScript respects the `@internal` when attempting to use the API from an external package. In this case, this utility code is simple enough to just copy over into this repository.