Add `enforcementLevel` to `PolicyPackArgs` (#205)
This is just a convenience for policy pack authors. Instead of having to define a default `enforcementLevel` value for each policy in the policy pack, this provides a place where such a default can be specified once. Setting it on any individual policies overrides the default value for those policies. If no `enforcementLevel` is set, defaults to "advisory".