Add wrappers for dict/list that raise on unknown values
This allows us to skip running a policy when it accesses an unknown value. When that happens, and advisory diagnostic is emitted. (Note: as a future improvement, we plan to create a new kind of warning event for these cases, as this isn't really an advisory violation -- this just matches the current behavior we have for Node.js policies.)