hacking around with running a policy
pipenv really doesn't want to work on my machine so I've just commented
out some python bits for now.

This is a minimal go program that can be used to test with "TestConfig"
that go works. I've got the engine running it! So it's just a case of
getting this all working SDK side now.