Add a secrets preserving proxy (Node.js)
This change preserves secrets round-tripping through policy
remediations. This is the Node.js side of the change, and Python
will come shortly. This works by preserving the secret nature of
secrets as they are read from property state. All values are
proxied, however, so that they can still be used as plain values.
The proxying will trap all writes to objects and ensure that
secretness is preserved on all properties that were secret on their
way in to the remediation. The serialization code then detects and
preserves secret values as state is returned from a remediation.