Switch to API Token for PyPi uploads (#322)
Username/Password authentication is no longer supported. This PR migrates to using API Tokens per https://pypi.org/help/#apitoken. We use `__token__` for the username and the org-wide `PYPI_API_TOKEN` secret for the password.

Also, fix a test due to missing function in `@pulumi/pulumi` that is now failing. A test that uses `_setFeatureSupport` to enable secrets serialization is now failing because this function no longer exists in `@pulumi/pulumi`. This change updates the test to call `setMocks` instead, which also enables secrets serialization.