Upgrade Go dependencies and make CI run on Go 1.21.x (#327)
We need to run on a more recent version of Go in order to address https://github.com/pulumi/pulumi-policy/pull/321. This PR updates our Go dependencies to the latest versions and updates CI to use Go 1.21.x.

Also deleted a `go.mod` and `go.sum` from the `./scripts` directory, as it's no longer needed. That directory used to have a go file, but it's long since been deleted.