Treat "remediate" stack policies as "mandatory"
Stack policies cannot be remediated (resource policies can). It's possible to configure a policy pack such that all or certain policies have a certain enforcement level. This means it's currently possible to specify an enforcement level of "remediate" for a stack policy even though stack policies do not support being remediated.

Currently, if a stack policy has a level of "remediate" and a violation is reported, a panic occurs in the CLI.

This commit addresses this by treating stack policies with a level of "remediate" as "mandatory", similar to how "remediate" resource policies are treated as "mandatory" if the policy is still in violation after a remediation has run.