[sdk/python] Allow editable installs without build step
This change makes it possible to install the Python SDK with `pip install -e sdk/python/lib` without having to first build the project and then install from the build directory with `pip install -e sdk/python/env/src`.

This helps speed up local development, but also makes it possible to install in this way from tests that require the version of the package to be a simple version like `1.0.0` and not `1.12.0a1721429785` like with the built package.