set fetch-depth: 0 for getting the repo
Even though `fetch-tags` claims in the docs to fetch all the tags, no
matter the `fetch-depth` setting, this does not seem to be true, as
pulumictl still can't find the tags for some reason. Also set
`fetch-depth: 0`, so we just clone the whole repo. This is still
cheaper than what we did before, where we did a shallow clone, and
then unshallowed the repo, making the server doing all the work to
generate the shallow clone, and then fetching everything anyway.