[nodejs] serializeProperty: Fix undefined/null handling (#400)
The Node.js Policy SDK has an internal modified copy of `@pulumi/pulumi`'s `serializeProperties` function (which has its own special handling for secret proxies in the Policy SDK), and there are tests that ensures the behavior of the copy matches that of `@pulumi/pulumi`.

As part of moving from using `require(...)` to regular TypeScript `import` statements to get type checking, the compile errors in the tests uncovered that the tests weren't actually confirming the same behavior in some cases.

In particular, the policy SDK is always serializing `undefined` as `null`, but that's not the same behavior as in `@pulumi/pulumi`. Only in the following circumstances should `undefined` be serialized as `null`:

1. `undefined` elements in an array
2. `undefined` value of a special secret object

This change fixes the implementation in the Policy SDK to match the behavior in `@pulumi/pulumi` and updates the tests accordingly so that they are actually confirming the results are the same.